add gates

+++ app/Http/Controllers/SparePartInShopController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Gate;
 use Auth;
 use App\SparePartInShop;
 use Illuminate\Http\Request;
@@ -13,6 +14,9 @@ class SparePartInShopController extends Controller
     }
 
     public function create(Request $request){
+        if(Gate::denies('create-spare-part-in-shop'))
+            return back();
+
         $this->validate($request, [
             'spare_part_id' => 'required|exists:spare_parts,id',
             'price' => 'required|integer'
@@ -28,6 +32,9 @@ class SparePartInShopController extends Controller
     }
 
     public function delete(SparePartInShop $sparePartInShop){
+        if(Gate::denies('modify-spare-part-in-shop'))
+            return back();
+
         $sparePartInShop->delete();
 
         return redirect()->route('profile');
